refactor(functions): remove deprecated namespaced API
