+++ packages/functions/__tests__/httpsCallable-timeout-native.test.ts
@@ -1,4 +1,4 @@
-import { afterAll, beforeAll, beforeEach, describe, expect, it, jest } from '@jest/globals';
+import { beforeEach, describe, expect, it, jest } from '@jest/globals';
 
 jest.mock('@react-native-firebase/app/dist/module/common', () => {
   const actualCommon = jest.requireActual(
@@ -11,7 +11,8 @@ jest.mock('@react-native-firebase/app/dist/module/common', () => {
   };
 });
 
-import { firebase } from '../lib';
+// @ts-ignore test
+import { getFunctions, httpsCallable } from '../lib';
 
 function timeoutFromNativeCall(nativeMock: jest.Mock): number {
   const call = nativeMock.mock.calls[0] as unknown[] | undefined;
@@ -24,22 +25,12 @@ function timeoutFromNativeCall(nativeMock: jest.Mock): number {
 describe('httpsCallable timeout units on native platforms', function () {
   let httpsCallableNative: jest.Mock;
 
-  beforeAll(function () {
-    // @ts-ignore test-only global
-    globalThis.RNFB_SILENCE_MODULAR_DEPRECATION_WARNINGS = true;
-  });
-
-  afterAll(function () {
-    // @ts-ignore test-only global
-    globalThis.RNFB_SILENCE_MODULAR_DEPRECATION_WARNINGS = false;
-  });
-
   beforeEach(function () {
     httpsCallableNative = jest
       .fn<(...args: unknown[]) => Promise<{ data: null }>>()
       .mockResolvedValue({ data: null });
 
-    const functions = firebase.app().functions();
+    const functions = getFunctions();
     // @ts-ignore test-only internal cache
     functions._nativeModule = {
       httpsCallable: (
@@ -53,7 +44,8 @@ describe('httpsCallable timeout units on native platforms', function () {
   });
 
   it('converts milliseconds to seconds for httpsCallable (Android/iOS)', async function () {
-    const runner = firebase.app().functions().httpsCallable('example', { timeout: 30000 });
+    const functions = getFunctions();
+    const runner = httpsCallable(functions, 'example', { timeout: 30000 });
     await runner();
 
     expect(httpsCallableNative).toHaveBeenCalledTimes(1);
@@ -61,9 +53,10 @@ describe('httpsCallable timeout units on native platforms', function () {
   });
 
   it('does not mutate a reused options object on native platforms', async function () {
+    const functions = getFunctions();
     const sharedOptions = { timeout: 30000 };
-    const runnerA = firebase.app().functions().httpsCallable('exampleA', sharedOptions);
-    const runnerB = firebase.app().functions().httpsCallable('exampleB', sharedOptions);
+    const runnerA = httpsCallable(functions, 'exampleA', sharedOptions);
+    const runnerB = httpsCallable(functions, 'exampleB', sharedOptions);
 
     await runnerA();
     await runnerB();
